Move bulk of stripe logic into 'billing' app
