+++ webapp/apps/billing/urls.py
@@ -2,5 +2,5 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('signup/', views.SignUp.as_view(), name='signup'),
+    path('webhook/', views.Webhook.as_view(), name='webhook'),
 ]
