+++ webapp/apps/users/tests/test_views.py
@@ -1,5 +1,3 @@
-import time
-
 import pytest
 
 from django.contrib.auth import get_user_model
@@ -12,14 +10,14 @@ class TestUsersViews():
     def test_get_login(self, client):
         resp = client.get('/users/signup/')
         assert resp.status_code == 200
-    
+
     def test_post_login(self, client, password):
         data = {
-            'csrfmiddlewaretoken': ['abc123'], 
-            'username': ['tester'], 
-            'email': ['tester@testing.ai'], 
-            'password1': [password], 
-            'password2': [password], 
+            'csrfmiddlewaretoken': ['abc123'],
+            'username': ['tester'],
+            'email': ['tester@testing.ai'],
+            'password1': [password],
+            'password2': [password],
             'stripeToken': ['tok_bypassPending']
         }
         resp = client.post('/users/signup/', data=data)
